updated test setup

+++ tickets/src/test/setup.ts
@@ -21,10 +21,12 @@ beforeAll ( async () => {
 
 beforeEach ( async () => {
   jest.clearAllMocks ();
-  const collections = await mongoose.connection.db!.collections ();
-  
-  for ( let collection of collections ) {
-    await collection.deleteMany ( {} );
+  if ( mongoose.connection.db ) {
+    const collections = await mongoose.connection.db.collections ();
+    
+    for ( let collection of collections ) {
+      await collection.deleteMany ( {} );
+    }
   }
 } );
